+++ lib/install.ps1
@@ -1231,14 +1231,16 @@ function persist_permission($manifest, $global) {
 # test if there are running processes
 function test_running_process($app, $global) {
     $processdir = appdir $app $global | Convert-Path
-    $running_processes = Get-Process | Where-Object { $_.Path -like "$processdir\*" }
+    $running_processes = Get-Process | Where-Object { $_.Path -like "$processdir\*" } | Out-String
 
     if ($running_processes) {
         if (get_config 'ignore_running_processes') {
-            warn "Application `"$app`" is still running. Scoop is configured to ignore this condition."
+            warn "The following instances of `"$app`" are still running. Scoop is configured to ignore this condition."
+            Write-Host $running_processes
             return $false
         } else {
-            error "Application `"$app`" is still running. Close all instances and try again."
+            error "The following instances of `"$app`" are still running. Close them and try again."
+            Write-Host $running_processes
             return $true
         }
     } else {