+++ libexec/scoop-hold.ps1
@@ -32,6 +32,12 @@ if ($global -and !(is_admin)) {
 $apps | ForEach-Object {
     $app = $_
 
+    if ($app -eq 'scoop') {
+        $hold_update_until = [System.DateTime]::Now.AddDays(1)
+        set_config 'hold_update_until' $hold_update_until.ToString('o') | Out-Null
+        success "$app is now held and might not be updated until $($hold_update_until.ToLocalTime())."
+        return
+    }
     if (!(installed $app $global)) {
         if ($global) {
             error "'$app' is not installed globally."